Add expanded footnote about comparison and Unicode
The original text was ambiguous (are comparisons
done by code point or code value?). This commit adds
a footnote so that those who want to find out more
can get an accurate definition of comparison in
Scala/Java. The idea of the footnote is to point out
that supplementary characters exist and may need to
be handled in a special way, while not going into the
details of how supplementary characters are encoded.
The linked Wikipedia page describes the details (and
also explains a little about Unicode and its history,
unlike the equivalent page on Swedish Wikipedia).