bank: Remove the State class
The State class is not particularly useful, since it
only is a very thin wrapper around a Queue. Queue is
also not a class that students have gotten familiar
with during the course.

I believe that the point of the State class is to
be able to have a map from dates to State objects,
but that kind of design with nested collections can
be hard to understand (and that design also happens
to waste a significant amount of memory), so I'm going
to make that design optional in the next commit.
Removing the State class is a bit of a necessity for
not making things too confusing for those who decide
to go with a different design.