bank: Clarify history-related instructions and relax requirements
The old instructions for the history of state changes
were confusing in several places, in particular because
they used the word "state" to refer to State objects
(which actually are changelogs rather than states) and
other things that are related to State objects but
aren't actually states. This commit rewrites parts of
the instructions in an attempt to clarify things.

This commit also removes the instructions' restrictions on
what data structure to use for the history, because the old
intended data structure was unnecessarily complicated in my
opinion. Students are still allowed to use the old intended
design if they want to (though the instructions don't point
them towards it at all), but I think it will be easier for
them (and as a bonus, the memory usage will be significantly
lower) if they just store a sequence of changes rather than
a map containing sequences of changes.

(To be honest, I'm not actually 100% sure if the old
instructions required the usage of the old intended data
structure, since the instructions were unclear. But
either way, after this commit, the instructions do not
require the usage of the old intended data structure :))