bank: Remove addAccount and removeAccount from spec
There are only two reasonable ways to implement these
two methods, and neither is good design:

1. Both methods create a BankEvent and call doEvent.
This would be reasonable behavior, but then why are
there such methods for creating/removing accounts but
not for depositing/withdrawing/transferring money?

2. Both methods modify the state directly without
creating a BankEvent. In this case, the methods must
not be called by anything other than doEvent, so they
should not be public.