Revert the rollback-related part of cf3c0d4
In cf3c0d4's version of the rollback example, it's hard to
understand what's going on unless the student has read and
understood the provided example file, which they likely haven't
done the first time they're reading through the instructions in
the compendium. Even if they have read it, they might have
forgotten the exact names used in it and might thus not realize
that the rollback example is based on the provided example file.
cf3c0d4 also missed part of the point of trying to deposit money
into an account that doesn't exist - it's supposed to demonstrate
that even though this action was possible before the rollback,
it isn't possible after the rollback. Because of these reasons,
this commit reverts the changes to the rollback example.

The very last example was reverted by cf3c0d4 for an unexplained
reason, but I suppose it could've been because of how incorrect
it was. This commit re-adds it with fixed values, because I
think it's useful to have several examples of what kind of text
should be displayed when a transaction fails.

However, the other parts of cf3c0d4 were good. While reverting
the changes mentioned above, I've made sure that these examples
now use the consistent IDs that cf3c0d4 went with.