bank: Add a menu option for displaying the history
A problem with the progression was that the student had to
implement storing the history relatively early despite
not having a good way to test it until late into the
project. Getting it right is non-trivial but important,
so this commit adds a requirement for making a menu option
that prints out the history, in order to facilitate testing.