Never generate three numbers in course codes
In 2017, the LTH course codes were changed to
always contain two numbers rather than three.

Note: I'm not sure if all of the letters A-Z can occur
in the fourth position, but there is a small number of
courses now that use letters other than the letters
ACFGLMNP that were used last year, so I thought I might
as well allow all letters in order to simplify things.