Adding three dots
I think i found the simplest solution and something that would still make people remember the example. You can always ask the class about what is missing and see if they figure it out. The example of gender as a binary thing is a flawed way to describe human expression, behaviour and interaction.