Fix outdated info in answer for exercise 7 task 5
In Scala 2.13, calling toArray on an Array does create a new Array.

There are still cases in Scala 2.13 where calling toXXX doesn't
create a new collection, but as far as I know this only happens
with immutable collections, so the fact that there are cases
where no new collection gets created is mostly irrelevant.