Use real name instead of username in version control example
In my experience, most people who do not wish to be pseudonymous
set git's user.name setting to their real name with spaces and
diacritics. There is no technical obstacle to doing so, there is
no expectation that the name is character-by-character identical
to whatever online username you are using, and it looks nicer.