update lect-w01-intro.tex to improve clarity
As "scala-cli" has "run" as a command, a beginner who has used the command "scala-cli run ." previous to reading this might mistakenly believe that "scala" also has the command "run". However, it doesn’t: "run" only refers to the main function in this example. It would therefore be more pedagogical if the main function was called something different (like "hi").