Refactor pitch classes to use an enum rather than an int for representaion, a list of names, and a map for parsing
I elected to change the names from using ascii sharp-signs (hash signs) to using (ascii) flat-signs (lowercase Bs).
This was done so the pitch class names are valid identifiers.
We could use backticks if we still wanted to use hash signs but I find this more readable.
In my experience, flats tend to be more common defaults in DAWs and in piano sheet music.

Care is to be taken if the code is to be refactored in the future since several implementations still relies on the ordinals of the variants.